Add proxy of RoomClimateRegulator service

+++ src/main/java/ge/edu/freeuni/sdp/service/scheduler/temperature/ScheduleManager.java
@@ -3,7 +3,6 @@ package ge.edu.freeuni.sdp.service.scheduler.temperature;
 import ge.edu.freeuni.sdp.service.scheduler.temperature.core.Schedule;
 import ge.edu.freeuni.sdp.service.scheduler.temperature.core.TemperatureEntry;
 
-import javax.print.attribute.standard.Media;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -20,7 +19,7 @@ public class ScheduleManager {
     @Path("/houses/{house_id}/floors/{floor_id}/?{start_time}&{end_time}")
     @GET
     public Schedule get(@PathParam("house_id") int houseId, @PathParam("floor_id") int floorId,
-                        @PathParam("start_time") int dateFrom, @PathParam("end_time") int dateTo ) {
+                        @PathParam("start_time") long dateFrom, @PathParam("end_time") long dateTo ) {
         return new Schedule(new Date(dateFrom), new Date(dateTo));
     }
